Rename SessionAware state machine interfacel to SessionListener and allow arbitrary listeners on server-side sessions.
